handling errors

+++ web-client/src/pages/Forget/index.js
@@ -1,68 +1,64 @@
-import React, { useState, useEffect } from "react";
-import "./styles.js";
+import React, { useState, useEffect } from "react"
+import "./styles.js"
 import { 
-      Container, 
-      ContainerLeft, 
-      FormContainer, 
-      InputEmail 
-      } from './styles'; 
+  Container, 
+  ContainerLeft, 
+  FormContainer, 
+  InputEmail 
+} from './styles' 
 
-import LogoImg from "../../assets/logo.png";
+import LogoImg from "../../assets/logo.png"
 
-import { Link, useHistory } from "react-router-dom";
-import axios from "axios";
-import api from "../../services/api";
+import { Link, useHistory } from "react-router-dom"
+import axios from "axios"
+import api from "../../services/api"
 
 
 
 
 //---Começo API---//
-require('dotenv/config');
+require('dotenv/config')
 
 export default function Forget() {
-  const history = useHistory();
+  const history = useHistory()
 
-  const [mail, setMail] = useState("");
-
-  const [user, setUser] = useState({});
+  const [mail, setMail] = useState('')
 
   function submit(e) {
-    e.preventDefault();
+    e.preventDefault()
+
+    if(mail === undefined || mail === '') return alert('Você deve preencher o campo para prosseguir!')
+    else if(!mail.includes('@') && !mail.includes('.')) return alert('E-mail inválido!')
 
     const body = {mail: mail}
 
     api.post('/forgot', body).then((res) => {
 
-      if(res.Error){
-        // temporário
-        console.log('erro');
-      }
-      localStorage.clear();
-      const mail_auth_token = res.data.auth_token;
+      if(res.Error) console.log(`Erro: ${res}`)
+      
+      localStorage.clear()
+
+      const mail_auth_token = res.data.auth_token
 
       setTimeout(() => { 
-
-        alert(res.data.message); //alert temporário - PELO AMOR DE DEUS, NÃO ESQUECER DE TIRAR!!!!!
-        goToLogin();
-
-      }, 3000);
+        alert(res.data.message) //alert temporário - PELO AMOR DE DEUS, NÃO ESQUECER DE TIRAR!!!!!
+      
+        goToLogin()
+      }, 3000)
 
     }).catch((e) => {
+      localStorage.clear()
+      console.log(e)
 
-      localStorage.clear();
-      console.log(e);
-      alert('Não foi possível Enviar o email.\nTente novamente mais tarde.');
+      alert(e)
 
-    });
+    })
   }
 
   function goToLogin() {
-    history.push('/');
+    history.push('/')
   }
 
-//---Final API---//
-
-//---Começo do Front-end---//
   return (
       <Container>
         
@@ -72,11 +68,11 @@ export default function Forget() {
              
         <FormContainer>
           <img src={LogoImg} alt="Logo" />
-          <h1>Digite Seu email Aqui!</h1>
-          <InputEmail placeholder="E-mail"/>
-          <a>Solicitar troca de senha</a>
-          <a href="/">Voltar para o login!</a>
+          <h1>Digite seu email aqui!</h1>
+          <InputEmail type="email" onChange={e => setMail(e.target.value)} placeholder="E-mail"/>
+          <a onClick={submit} type='submit' >Solicitar troca de senha</a>
+          <a href="/">Voltar para o login</a>
         </FormContainer>
       </Container>
-  );
+  )
 }
